Improved/Alternate Implementation of Array List
Element destructors and comparators for `badalist` are now provided to
the calls which need them as arguments, rather than storing them in the
data structure, allowing the caller to decide at call time whether or
not they would like to destroy the element(s) or how they would like to
compare them. Use `realloc` in `alist_resize`. Allocate only the space
the caller asked for, and no more.