Bugfix Iterator Movement Operators
Fix `liter_push_back/front` so that both call `va_end` and clean up any
resources allocated, being sure not to clobber the `out` parameter if it
is the same an the `iter` parameter. Fix `liter_advance/prev/next` to
allow them to move to or from, but not over, the anchor node.