Improve ArrayList
Add stack functions for `ArrayList`s. Improve unit tests for
`ArrayList`s. Return `-1` on failure for all `ArrayList` functions which
return a status code. Change `ArrayList` functions to operate in terms
of the size and capacity of the list, rather than the size of the list
and the number of non-`NULL` elements currently stored in the list.